+++ apx_sdk/src/url.rs
@@ -21,7 +21,7 @@ pub const GATEWAY_PATH_PREFIX: &str = "/.well-known/apgateway/";
 pub struct ObjectIdError(pub &'static str);
 
 // TODO: FEP-EF61: rename to CanonicalUrl
-#[derive(Clone)]
+#[derive(Clone, PartialEq)]
 pub enum Url {
     Http(HttpUrl),
     Ap(ApUrl),